pyproject.toml
[project]
name = "winter-daily-digest"
version = "0.1.0"
description = "Automated daily digest of aespa Winter's Twitter updates, delivered via Telegram"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "python-telegram-bot>=22.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]
